refactor: domain 종속 컴포넌트 디렉토리 이관
