+++ src/components/domain/CardNumberInput/index.tsx
@@ -1,10 +1,8 @@
 import { memo, useEffect, useState } from 'react';
 
+import { useFormContext } from '@/components/common/Form/FormContext';
 import { InputContainer } from '@/components/UI';
 import { useNumberKeyInterceptor } from '@/hooks/useNumberKeyInterceptor';
-
-import { useFormContext } from '../common/Form/FormContext';
-
 type Props = {
   onChange: <T>(value: T) => void;
   dirtyState: boolean;
