+++ src/components/domain/CardOwnerInput/index.tsx
@@ -1,8 +1,7 @@
 import { memo, useEffect, useState } from 'react';
 
 import { InputContainer } from '@/components/UI';
-
-import { useFormContext } from '../common/Form/FormContext';
+import { useFormContext } from '@/components/common/Form/FormContext';
 
 type Props = {
   onChange: <T>(value: T) => void;
